+++ src/datalevin/server.clj
@@ -2081,11 +2081,14 @@
   (let [timeout (.-idle-timeout server)
         clients (.-clients server)]
     (doseq [[client-id session] clients
-            :let                [{:keys [last-active]} session]
-            :when               last-active]
-      (when (< ^long timeout
-               (- (System/currentTimeMillis) ^long last-active))
-        (disconnect-client* server client-id)))))
+            :let                [{:keys [last-active]} session]]
+      (if last-active
+        (when (< ^long timeout
+                 (- (System/currentTimeMillis) ^long last-active))
+          (disconnect-client* server client-id))
+        ;; migrate old sessions that don't have last-active
+        (update-client server client-id
+                       #(assoc % :last-active (System/currentTimeMillis)))))))
 
 (defn- event-loop
   [^Server server]